+++ packages/plugin-table/web/src/toolbar/buttons/BorderButton.tsx
@@ -4,7 +4,8 @@ import { RicosContext, ModalContext } from 'ricos-context';
 import { BorderIcon } from '../../icons';
 import type { IToolbarItem } from 'ricos-types';
 import { ToggleButton, DropdownButton } from 'wix-rich-content-toolbars-ui';
-import { TABLE_BUTTONS_MODALS_ID } from '../../consts';
+import { BORDER_TYPES, CATEGORY, TABLE_BUTTONS_MODALS_ID, TABLE_COLOR_PICKER } from '../../consts';
+import type { bordersType } from '../../types';
 
 interface Props {
   toolbarItem: IToolbarItem;
@@ -14,24 +15,21 @@ interface Props {
 const BorderButton: FC<Props> = ({ dataHook, toolbarItem }) => {
   const { t, isMobile } = useContext(RicosContext) || {};
   const modalService = useContext(ModalContext) || {};
-
   const [referenceElement, setReferenceElement] = useState<HTMLDivElement | null>(null);
-
-  const isMultipleSelection = true; // TODO: selectedCells.length > 1
-
+  const currentColor = toolbarItem.attributes?.cellBorderColor || 'unset';
+  const isMultipleSelection = !(toolbarItem.attributes?.selectedCategory === CATEGORY.CELL_BORDER);
+  const closeModal = () => modalService.closeModal(TABLE_BUTTONS_MODALS_ID.BORDER);
   const onColorSelect = (data: { borders?: string; outsideBorders?: string }) => {
     toolbarItem.commands?.click(data);
-    closeModal();
+    modalService.closeModal(TABLE_COLOR_PICKER);
   };
 
-  const closeModal = () => modalService.closeModal(TABLE_BUTTONS_MODALS_ID.BORDER);
-
-  const openBorderPanel = () => {
+  const toggleBorderPanel = () => {
     const isModalOpen = modalService.isModalOpen(TABLE_BUTTONS_MODALS_ID.BORDER);
     !isModalOpen
       ? modalService.openModal(TABLE_BUTTONS_MODALS_ID.BORDER, {
           componentProps: {
-            onClick: onColorSelect,
+            onClick: openColorPickerPanel,
             closeModal,
           },
           layout: isMobile ? 'drawer' : 'toolbar',
@@ -40,18 +38,37 @@ const BorderButton: FC<Props> = ({ dataHook, toolbarItem }) => {
       : closeModal();
   };
 
+  const openColorPickerPanel = (borderType: bordersType) => {
+    const isModalOpen = modalService.isModalOpen(TABLE_COLOR_PICKER);
+    const closeModal = () => modalService.closeModal(TABLE_COLOR_PICKER);
+    !isModalOpen
+      ? modalService?.openModal(TABLE_COLOR_PICKER, {
+          componentProps: {
+            onChange: color => onColorSelect({ [borderType]: color }),
+            closeModal,
+            currentColor,
+            resetColor: () => onColorSelect({ [borderType]: 'unset' }),
+          },
+          positioning: { placement: 'bottom', referenceElement },
+          layout: isMobile ? 'drawer' : 'toolbar',
+        })
+      : closeModal();
+  };
+
   return !isMultipleSelection ? (
-    <ToggleButton
-      Icon={BorderIcon}
-      onClick={color => onColorSelect({ borders: color })}
-      dataHook={dataHook}
-      tooltip={t('ButtonModal_Border_Color')}
-    />
+    <div ref={setReferenceElement}>
+      <ToggleButton
+        Icon={BorderIcon}
+        onClick={() => openColorPickerPanel(BORDER_TYPES.borders)}
+        dataHook={dataHook}
+        tooltip={t('ButtonModal_Border_Color')}
+      />
+    </div>
   ) : (
     <DropdownButton
       dataHook={dataHook}
       active={modalService.isModalOpen(TABLE_BUTTONS_MODALS_ID.BORDER)}
-      onClick={openBorderPanel}
+      onClick={toggleBorderPanel}
       setRef={setReferenceElement}
       Icon={BorderIcon}
       tooltip={t('ButtonModal_Border_Color')}